Fix: Link logo to site by overridng theme bug
This fix works by overriding a buggy template block in the
Terminimal theme where part of the path to `logo_home_link`
is missing.

The Terminimal theme has a bug where adding a `logo_home_link`
under `[extra]` in `config.toml` yields no change. Instead,
it continues to point to `base_url`. This bug is caused by
a missing `.extra` when accessing `home_link` with Tera.